refactor: move samplesDirectory to a utility file for reuse
RISDEV-5805
