+++ frontend/e2e/announcements-upload.spec.ts
@@ -4,7 +4,7 @@ import { expect, test } from "@playwright/test"
 import fs from "fs"
 import path from "node:path"
 import { uploadAmendingLaw } from "@e2e/utils/upload-with.force"
-import { samplesDirectory } from "@e2e/globalSetup/global-setup"
+import { samplesDirectory } from "@e2e/utils/samples-directory"
 
 test(
   "navigates to the upload page",
